fix token in setup
